requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
pydantic-settings==2.6.1
